# Refactor Knn Search Results to use TopDocs