--- a/src/main/java/org/opensearch/knn/index/query/KNNWeight.java
+++ b/src/main/java/org/opensearch/knn/index/query/KNNWeight.java
@@ -14,8 +14,11 @@ import org.apache.lucene.index.SegmentReader;
 import org.apache.lucene.search.DocIdSetIterator;
 import org.apache.lucene.search.Explanation;
 import org.apache.lucene.search.FilteredDocIdSetIterator;
+import org.apache.lucene.search.ScoreDoc;
 import org.apache.lucene.search.Scorer;
 import org.apache.lucene.search.ScorerSupplier;
+import org.apache.lucene.search.TopDocs;
+import org.apache.lucene.search.TotalHits;
 import org.apache.lucene.search.Weight;
 import org.apache.lucene.util.BitSet;
 import org.apache.lucene.util.BitSetIterator;
@@ -41,9 +44,7 @@ import org.opensearch.knn.plugin.stats.KNNCounter;
 
 import java.io.IOException;
 import java.util.Arrays;
-import java.util.Collections;
 import java.util.List;
-import java.util.Map;
 
 import static org.opensearch.knn.common.KNNConstants.KNN_ENGINE;
 import static org.opensearch.knn.common.KNNConstants.MODEL_ID;
@@ -70,6 +71,7 @@ import static org.opensearch.knn.common.KNNConstants.VECTOR_DATA_TYPE_FIELD;
  */
 @Log4j2
 public abstract class KNNWeight extends Weight {
+    protected static final TopDocs EMPTY_TOPDOCS = new TopDocs(new TotalHits(0, TotalHits.Relation.EQUAL_TO), new ScoreDoc[0]);
     private static ModelDao modelDao;
     private static ExactSearcher DEFAULT_EXACT_SEARCHER;
 
@@ -122,7 +124,7 @@ public abstract class KNNWeight extends Weight {
     public Explanation explain(LeafReaderContext context, int doc, float score) {
         knnQuery.setExplain(true);
         try {
-            final KNNScorer knnScorer = getOrCreateKnnScorer(context);
+            final Scorer knnScorer = getOrCreateKnnScorer(context);
             // calculate score only when its 0 as for disk-based search,
             // score will be passed from the caller and there is no need to re-compute the score
             if (score == 0) {
@@ -242,9 +244,9 @@ public abstract class KNNWeight extends Weight {
         return sb.toString();
     }
 
-    private KNNScorer getOrCreateKnnScorer(LeafReaderContext context) throws IOException {
+    private Scorer getOrCreateKnnScorer(LeafReaderContext context) throws IOException {
         // First try to get the cached scorer
-        KNNScorer scorer = knnExplanation.getKnnScorer(context);
+        Scorer scorer = knnExplanation.getKnnScorer(context);
 
         // If no cached scorer exists, create and cache a new one
         if (scorer == null) {
@@ -255,7 +257,7 @@ public abstract class KNNWeight extends Weight {
         return scorer;
     }
 
-    private float getKnnScore(KNNScorer knnScorer, int doc) throws IOException {
+    private float getKnnScore(Scorer knnScorer, int doc) throws IOException {
         return (knnScorer.iterator().advance(doc) == doc) ? knnScorer.score() : 0;
     }
 
@@ -266,13 +268,12 @@ public abstract class KNNWeight extends Weight {
 
             @Override
             public Scorer get(long leadCost) throws IOException {
-                final Map<Integer, Float> docIdToScoreMap = searchLeaf(context, knnQuery.getK()).getResult();
-                cost = docIdToScoreMap.size();
-                if (docIdToScoreMap.isEmpty()) {
+                final TopDocs topDocs = searchLeaf(context, knnQuery.getK()).getResult();
+                cost = topDocs.scoreDocs.length;
+                if (cost == 0) {
                     return KNNScorer.emptyScorer();
                 }
-                final int maxDoc = Collections.max(docIdToScoreMap.keySet()) + 1;
-                return new KNNScorer(ResultUtil.resultMapToDocIds(docIdToScoreMap, maxDoc), docIdToScoreMap, boost);
+                return new KNNScorer(topDocs, boost);
             }
 
             @Override
@@ -317,7 +318,7 @@ public abstract class KNNWeight extends Weight {
          * This improves the recall.
          */
         if (isFilteredExactSearchPreferred(cardinality)) {
-            Map<Integer, Float> result = doExactSearch(context, new BitSetIterator(filterBitSet, cardinality), cardinality, k);
+            TopDocs result = doExactSearch(context, new BitSetIterator(filterBitSet, cardinality), cardinality, k);
             return new PerLeafResult(filterWeight == null ? null : filterBitSet, result);
         }
 
@@ -328,20 +329,20 @@ public abstract class KNNWeight extends Weight {
         final BitSet annFilter = (filterWeight != null && cardinality == maxDoc) ? null : filterBitSet;
 
         StopWatch annStopWatch = startStopWatch();
-        final Map<Integer, Float> docIdsToScoreMap = approximateSearch(context, annFilter, cardinality, k);
+        final TopDocs topDocs = approximateSearch(context, annFilter, cardinality, k);
         stopStopWatchAndLog(annStopWatch, "ANN search", segmentName);
         if (knnQuery.isExplain()) {
-            knnExplanation.addLeafResult(context.id(), docIdsToScoreMap.size());
+            knnExplanation.addLeafResult(context.id(), topDocs.scoreDocs.length);
         }
         // See whether we have to perform exact search based on approx search results
         // This is required if there are no native engine files or if approximate search returned
         // results less than K, though we have more than k filtered docs
-        if (isExactSearchRequire(context, cardinality, docIdsToScoreMap.size())) {
+        if (isExactSearchRequire(context, cardinality, topDocs.scoreDocs.length)) {
             final BitSetIterator docs = filterWeight != null ? new BitSetIterator(filterBitSet, cardinality) : null;
-            Map<Integer, Float> result = doExactSearch(context, docs, cardinality, k);
+            TopDocs result = doExactSearch(context, docs, cardinality, k);
             return new PerLeafResult(filterWeight == null ? null : filterBitSet, result);
         }
-        return new PerLeafResult(filterWeight == null ? null : filterBitSet, docIdsToScoreMap);
+        return new PerLeafResult(filterWeight == null ? null : filterBitSet, topDocs);
     }
 
     private void stopStopWatchAndLog(@Nullable final StopWatch stopWatch, final String prefixMessage, String segmentName) {
@@ -383,7 +384,7 @@ public abstract class KNNWeight extends Weight {
         return BitSet.of(filterIterator, maxDoc);
     }
 
-    private Map<Integer, Float> doExactSearch(
+    private TopDocs doExactSearch(
         final LeafReaderContext context,
         final DocIdSetIterator acceptedDocs,
         final long numberOfAcceptedDocs,
@@ -401,18 +402,14 @@ public abstract class KNNWeight extends Weight {
         return exactSearch(context, exactSearcherContextBuilder.build());
     }
 
-    private Map<Integer, Float> approximateSearch(
-        final LeafReaderContext context,
-        final BitSet filterIdsBitSet,
-        final int cardinality,
-        final int k
-    ) throws IOException {
+    private TopDocs approximateSearch(final LeafReaderContext context, final BitSet filterIdsBitSet, final int cardinality, final int k)
+        throws IOException {
         final SegmentReader reader = Lucene.segmentReader(context.reader());
         FieldInfo fieldInfo = FieldInfoExtractor.getFieldInfo(reader, knnQuery.getField());
 
         if (fieldInfo == null) {
             log.debug("[KNN] Field info not found for {}:{}", knnQuery.getField(), reader.getSegmentName());
-            return Collections.emptyMap();
+            return EMPTY_TOPDOCS;
         }
 
         KNNEngine knnEngine;
@@ -450,14 +447,14 @@ public abstract class KNNWeight extends Weight {
         List<String> engineFiles = KNNCodecUtil.getEngineFiles(knnEngine.getExtension(), knnQuery.getField(), reader.getSegmentInfo().info);
         if (engineFiles.isEmpty()) {
             log.debug("[KNN] No native engine files found for field {} for segment {}", knnQuery.getField(), reader.getSegmentName());
-            return Collections.emptyMap();
+            return EMPTY_TOPDOCS;
         }
 
         // TODO: Change type of vector once more quantization methods are supported
         final byte[] quantizedVector = SegmentLevelQuantizationUtil.quantizeVector(knnQuery.getQueryVector(), segmentLevelQuantizationInfo);
 
         KNNCounter.GRAPH_QUERY_REQUESTS.increment();
-        final Map<Integer, Float> results = doANNSearch(
+        final TopDocs results = doANNSearch(
             context,
             reader,
             fieldInfo,
@@ -471,9 +468,9 @@ public abstract class KNNWeight extends Weight {
             k
         );
 
-        if (results.isEmpty()) {
+        if (results.scoreDocs.length == 0) {
             log.debug("[KNN] Query yielded 0 results");
-            return Collections.emptyMap();
+            return EMPTY_TOPDOCS;
         }
 
         return results;
@@ -496,7 +493,7 @@ public abstract class KNNWeight extends Weight {
      * @return A table maps document id to its Lucene score.
      * @throws IOException
      */
-    abstract protected Map<Integer, Float> doANNSearch(
+    abstract protected TopDocs doANNSearch(
         final LeafReaderContext context,
         final SegmentReader reader,
         final FieldInfo fieldInfo,
@@ -522,17 +519,27 @@ public abstract class KNNWeight extends Weight {
         }
     }
 
+    protected void addExplainIfRequired(final TopDocs results, final KNNEngine knnEngine, final SpaceType spaceType) {
+        if (knnQuery.isExplain()) {
+            Arrays.stream(results.scoreDocs).forEach(result -> {
+                if (KNNEngine.FAISS.getName().equals(knnEngine.getName()) && SpaceType.INNER_PRODUCT.equals(spaceType)) {
+                    knnExplanation.addRawScore(result.doc, -1 * result.score);
+                } else {
+                    knnExplanation.addRawScore(result.doc, result.score);
+                }
+            });
+        }
+    }
+
     /**
      * Execute exact search for the given matched doc ids and return the results as a map of docId to score.
      * @return Map of docId to score for the exact search results.
      * @throws IOException If an error occurs during the search.
      */
-    public Map<Integer, Float> exactSearch(
-        final LeafReaderContext leafReaderContext,
-        final ExactSearcher.ExactSearcherContext exactSearcherContext
-    ) throws IOException {
+    public TopDocs exactSearch(final LeafReaderContext leafReaderContext, final ExactSearcher.ExactSearcherContext exactSearcherContext)
+        throws IOException {
         StopWatch stopWatch = startStopWatch();
-        Map<Integer, Float> exactSearchResults = exactSearcher.searchLeaf(leafReaderContext, exactSearcherContext);
+        TopDocs exactSearchResults = exactSearcher.searchLeaf(leafReaderContext, exactSearcherContext);
         final SegmentReader reader = Lucene.segmentReader(leafReaderContext.reader());
         stopStopWatchAndLog(stopWatch, "Exact search", reader.getSegmentName());
         return exactSearchResults;