@VijayanB 2025-06-06
@shatejas 

Configuration:
Dataset: Cohere 1M
Data Node : 1

Benchmark results




Metrics | baseline | PR #2727
-- | -- | --
Segment Count | 30 | 30
P50 Service Time | 15.78 ms | 13.83ms
P90 Seervice Time | 19.03ms | 17.11ms
P99 Service Time | 23.48ms | 21.43ms
Recall@100 | 0.99 | 0.99
Recall@1 | 1 | 1