@VijayanB 2025-06-10
@shatejas 

Full exact search benchmark


Metrics | baseline | PR #2727
-- | -- | --
Segment Count | 32 | 32
Doc Count | 1M | 1M
Shard Count | 3 | 3
Node Count | 1 | 1
P50 Service Time | 176.814 | 175.448
P90 Seervice Time | 184.251 | 180.809
P99 Service Time | 189.225 | 185.591
error rate | 0 | 0
Recall@100 | 1 | 1
Recall@1 | 1 | 1